refactor(app): server render landing page
